refactor: Extract useEnsuredControl hook
